Fix memory corruption issue
When calling ParseUnsignedLong in BattleShip_UI_Read_Menu, the address of uint chosen_option was cast to unsigned long. This meant ParseUnsignedLong deferenced memory past the size of uint and corrupted chosen_option_str.

+++ root/battleship/include/battleship/util.h
@@ -11,7 +11,7 @@
 #include "battleship/commondefs.h"
 
 String_t TrimStr(String_t str, size_t str_size);
-bool ParseUnsignedLong(const char *str, unsigned long *val);
+bool ParseUnsignedLong(const char * str, uint *val);
 int CalcMax(const int *data, uint n);
 int CalcNumWidth(int n);
 int RepeatChar(String_t line, size_t size_line, char elem);
